requirements/base.txt
fastapi==0.111.0
uvicorn==0.18.2
python-dotenv==0.20.0
pymongo==4.3.3
motor==3.1.2
requests==2.32.4
password-validator==1.0
email-validator==2.0.0
python-dateutil==2.8.2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pandas==2.0.0
itsdangerous==2.1.2
fastapi-sso==0.7.2
user-agents==2.2.0
resend==2.11.0
slowapi==0.1.9